Fixed error where I used ints and not "long long" values for the cursor's. Note I have not used MGTwitterEngineID as the cursors need to be signed, not unsigned.
I have also added some error checking in the getFollowerIDs & getFriendIDs; a cursor of 0 is invalid, as is a username name of "".

I have also changed the format of the request to match the other API request methods

+++ MGTwitterEngine.h
@@ -184,8 +184,8 @@
 
 
 // Social Graph methods
-- (NSString *)getFriendIDsFor:(NSString *)username startingFromCursor:(int)cursor; // friends/ids
-- (NSString *)getFollowerIDsFor:(NSString *)username startingFromCursor:(int)cursor; // followers/ids
+- (NSString *)getFriendIDsFor:(NSString *)username startingFromCursor:(long long)cursor; // friends/ids
+- (NSString *)getFollowerIDsFor:(NSString *)username startingFromCursor:(long long)cursor; // followers/ids
 
 
 #pragma mark Search API methods
